+++ src/services/LogTrait.php
@@ -5,6 +5,6 @@ namespace crmpbx\commutator\services;
 trait LogTrait
 {
     public string $logServiceAddress = '';
-    public int $logServiceTimeout = 1;
+    public float $logServiceTimeout = 1;
     public string $logServiceAccessToken = '';
 }
