+++ src/services/PipedriveTrait.php
@@ -5,6 +5,6 @@ namespace crmpbx\commutator\services;
 trait PipedriveTrait
 {
     public string $pipedriveServiceAddress = '';
-    public int $pipedriveServiceTimeout = 1;
+    public float $pipedriveServiceTimeout = 1;
     public string $pipedriveTestServiceAddress = '';
 }
